requirements.txt
pillow
mcpi
pyTelegramBotApi
opencv-python